Refactor config generation
The old approach was to label nodes in the halconfig as belonging to certain
spinnaker components, and then appending that node to that component's config
when it was written. The benefit of that approach was it made it easy to
connect the halconfig to the spinnaker component's config, but it made it
impossible to generate more complicated config, and forced the halconfig model
to depend on the config generation service.

I've rewritten it so each service decides how to generate its own config.