fix(cli/accounts): Allows specifying environments for accounts
The `credentials.challengeDestructiveActionsEnvironments` setting in
clouddriver controls a few features in deck. This setting is based off
the configured environment for an account. Adding environment to the
add and edit account commands allows use of halyard to take advantage of
those deck features.

Moving the environment flag up from the Openstack subclasses makes it
available to all account types.