fix(install): Halyard startup should use a login shell
On Ubuntu 18.04 on GCE, gcloud is installed by snap in /snap/bin.
There is an entry to add /snap/bin to the path in /etc/profile.d
but this only runs for login shells, so if halyard runs in a
non-login shell it is unable to find gcloud. Rather than address
the specific case, we should start the halyard daemon in a login
shell.