refactor(core): Push getChildren to Node base class
In most cases, we want the children of a given node to be
any declared fields that are also nodes; many nodes currently
implement getChildren identically via a call to makeReflectiveIterator.
Even in cases where getChildren is defined as an empty iterator or
as an explicit iteration, it is usually just an explicit expression
of what makeReflectiveIterator would produce.

This change implements getChildren in the base class as a call to
makeReflectiveIterator. It does not remove overrides that explicitly
define their own iterator (even if that iterator is identical to that
generated by makeReflectiveIterator).