refactor(controller): Refactor common logic in controller GET handling
Replace the duplicated code in the controllers with calls to the new
GenericGetRequest class.  This should make it conceptually easier
to add new controllers for halyard settings, as it's now only
necessary to specify the getter, validator, and description of the
task, and the common logic is handled by GenericGetRequest.