fix(install): Move install script back to old path
The install script was moved to the 'linux' directory instead of
'debian' after adding support for arch linux.  While this is
technically more correct, the docs (and potentially scripts/automation)
point to the old location. Move the script back to the old location. At
some point in the future we can move it to the 'linux' directory but
we'll need to figure out any dependencies (docs, automation) first.