fix(build): Fix aws-iam-authenticator in docker container (#1278) (#1279)
Updating kubectl to 1.14 broke the link to download the
aws-iam-authenticator.  The highest currently-published version
is for kubectl 1.12.7.

Let's downgrade kubectl to 1.12.7 (which is still an upgrade from
the 1.10.3 it was in the prior version of halyard). In general as
I think of it more, we should probably try to keep this as N-1 or
N-2 of the newest kubectl version to try to have the most overlap
with where users are.