fix(install): Fix installation on debian-based systems
The call to 'which systemd-users' returns an exit status >0 when
it fails.  As we're running under 'set -e' this causes the script
to immediately exit in failure.

Bash does not exit in failure when the command that produces an
exit value >0 is part of an if condition.  That's actually what
we want here; just run the command directly in the 'if'.