fix(ci): Fix error casting CI class from daemon
It turns out that removing the objectMapper from the halyard-cli
commands broke deserializing the result from the daemon. We'll need
to continue to conver the result to the expected class.

Instead of adding back reflection, make a small enum that keeps track
of the classes relevant for each CI system, and call that enum to
do the mapping.