fix(kubernetes): Added the missing support for the HTTPS scheme to Kubernetes httpGet readiness probe
Setting kubernetes.useExecHealthCheck: false as a custom service setting
for gate configures the Kubernetes readiness check to use the built-in
httpGet functionality instead of using the external wget command. We
found that the HTTPS scheme is not specified when the endpoints use TLS
thereby resulting in failed readiness probes. This commit adds the
missing support for the HTTPS scheme.