feat(k8s/deploy): Add custom health check to service-settings.
Need the ability to override the default commands created and populated
in the Deployments readinessprobe. In our use case mTLS has been enabled
and we need a way to modify the readiness probe in each service so that
you can pass in the client certificate.

By adding this functionality Halyard can support a much wider array of potential
configurations for health checks for cases where a simple check of /health might not
be desirable.