refactor(deck): remove unneeded env vars from DeckDockerProfileFactory
When building Deck as a container, several environment variables are required when using SSL. However, the AUTH_ENABLED and FIAT_ENABLED environment variables are only read from Deck's settings.js when doing local development. When building Deck as a container, auth and fiat configs are read from the features configuration block, not environment variables, so we can safely remove this code.