fix(cli): fix validation of main parameters that specify arity
https://github.com/spinnaker/halyard/pull/1894 brought in version 1.75 of
com.beust:jcommander where we previously had 1.71.  1.75 includes
https://github.com/cbeust/jcommander/commit/192ec8695636e6793dfbebd8aaab2f1d467b922e which
adds additional arity validation, which causes halyard commands with main parameters that
specify an arity to fail.  I suspect it's a bug in jcommander, but...

1.75 also includes
https://github.com/cbeust/jcommander/commit/d51a5b722331c2b69bbf86c238a14a9a396c3b44 which
allows us to replace List<String> and arity of 1 with String, simplifying our code.  It
would be nice for jcommander's validation to handle the missing case, but this is still
simpler than what we had before, and works.

Bumping to 1.81 at the same time to stay up to date, and to demonstrate that specifying
arity = 1 still doesn't work with 1.81, but these fixes do.